Admin server works with session id only admins can manipulate books

+++ src/main/java/com/example/online_library/library/admin/LibraryAdminService.java
@@ -2,8 +2,8 @@ package com.example.online_library.library.admin;
 
 import com.example.online_library.exceptions.AdminAccessDeniedException;
 import com.example.online_library.login.encryptUserSession.EncryptionUtils;
-import com.example.online_library.mapper.mappers.LibraryRequestMapper;
 import com.example.online_library.mapper.dto.LibraryRequestDto;
+import com.example.online_library.mapper.mappers.LibraryRequestMapper;
 import com.example.online_library.models.appuser.UserRole;
 import com.example.online_library.models.appuser.UserService;
 import com.example.online_library.models.book.Book;
@@ -25,15 +25,12 @@ public class LibraryAdminService {
     private final UserService userService;
     private final LibraryRequestMapper libraryRequestMapper;
     private static final String SESSION_NAME = "MY_SESSION_ID";
+    private static final UserRole ADMIN = UserRole.ADMIN;
 
-    private void checkUserOrThrowException(HttpServletRequest httpServletRequest) {
-        System.out.println("HttpServletRequest: " + httpServletRequest.getSession());
 
-        String header = httpServletRequest.getHeader("Cookie");
-        System.out.println("Header 1: " + header);
-
+    private void checkSessionForUserExistingAndAdminRole(HttpServletRequest httpServletRequest) throws RuntimeException {
         Cookie[] cookies = httpServletRequest.getCookies();
-        System.out.println("Cookies[] 2: " + cookies);
+        System.out.println("Cookies[]: " + cookies);
 
         if (cookies == null) {
             throw new SessionException("Session cookie not found");
@@ -50,24 +47,30 @@ public class LibraryAdminService {
         return SESSION_NAME.equals(cookie.getName());
     }
 
-    private void handleSessionCookie(Cookie cookie) {
+    private void handleSessionCookie(Cookie cookie) throws RuntimeException{
         try {
             SecretKey secretKey = EncryptionUtils.generateSecretKey();
             String decryptedSessionData = decryptSessionCookie(cookie, secretKey);
 
-            if (decryptedSessionData.isBlank()) {
-                throw new SessionException("Session cookie not found");
-            }
+            System.out.println("Decrypted session: " + decryptedSessionData);
 
             String[] sessionParts = decryptedSessionData.split(":");
             String email = (sessionParts.length > 1) ? sessionParts[1] : null;
             String role = (sessionParts.length > 2) ? sessionParts[2] : null;
 
-            if (!userService.findUserByEmailAndRoleAdmin(email, UserRole.valueOf(role))) {
-                throw new AdminAccessDeniedException("User not found, or user is not an administrator");
+            System.out.println("Email: " + email);
+            System.out.println("Role: " + role);
+
+            if (!ADMIN.equals(UserRole.valueOf(role))) {
+                throw new AdminAccessDeniedException("User is not an administrator");
+            }
+
+            if (!userService.findUserByEmailAndRole(email, ADMIN)) {
+                throw new AdminAccessDeniedException("User not found or not an administrator");
             }
         } catch (Exception e) {
             e.printStackTrace();
+            throw new RuntimeException("Error handling session cookie", e);
         }
     }
 
@@ -78,28 +81,49 @@ public class LibraryAdminService {
 
 
     public Long getBookId(String title, String author, HttpServletRequest httpServletRequest) {
-        checkUserOrThrowException(httpServletRequest);
+
+        try {
+        checkSessionForUserExistingAndAdminRole(httpServletRequest);
 
         return bookAdminService.findBookId(title, author);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException("User does not have permission to add a book", e);
+        }
     }
 
     public void includeNewBookToLibrary(LibraryRequestDto request, HttpServletRequest httpServletRequest) {
-        checkUserOrThrowException(httpServletRequest);
+        try {
+            checkSessionForUserExistingAndAdminRole(httpServletRequest);
 
-        Book book = libraryRequestMapper.libraryRequestDtoToBook(request);
-        bookAdminService.addNewBook(book);
+            Book book = libraryRequestMapper.libraryRequestDtoToBook(request);
+            bookAdminService.addNewBook(book);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException("User does not have permission to add a book", e);
+        }
     }
 
     public void changeExistingBookInform(Long book_id, LibraryRequestDto request, HttpServletRequest httpServletRequest) {
-        checkUserOrThrowException(httpServletRequest);
+        try {
+            checkSessionForUserExistingAndAdminRole(httpServletRequest);
 
-        Book book = libraryRequestMapper.libraryRequestDtoToBook(request);
-        bookAdminService.updateExistingBook(book_id, book);
+            Book book = libraryRequestMapper.libraryRequestDtoToBook(request);
+            bookAdminService.updateExistingBook(book_id, book);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException("User does not have permission to add a book", e);
+        }
     }
 
     public void deleteBookFromLibrary(Long book_id, HttpServletRequest httpServletRequest) {
-        checkUserOrThrowException(httpServletRequest);
+        try {
+            checkSessionForUserExistingAndAdminRole(httpServletRequest);
 
-        bookAdminService.deleteBook(book_id);
+            bookAdminService.deleteBook(book_id);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException("User does not have permission to add a book", e);
+        }
     }
 }
